Use 64 bit counters to keep track of bytes sent/received from the virtual network interface.

+++ src/dummy/device.c
@@ -27,8 +27,8 @@ char *device = "dummy";
 char *iface = "dummy";
 static char *device_info = "dummy device";
 
-static int device_total_in = 0;
-static int device_total_out = 0;
+static uint64_t device_total_in = 0;
+static uint64_t device_total_out = 0;
 
 bool setup_device(void) {
 	logger(LOG_INFO, "%s (%s) is a %s", device, iface, device_info);
@@ -49,6 +49,6 @@ bool write_packet(vpn_packet_t *packet) {
 
 void dump_device_stats(void) {
 	logger(LOG_DEBUG, "Statistics for %s %s:", device_info, device);
-	logger(LOG_DEBUG, " total bytes in:  %10d", device_total_in);
-	logger(LOG_DEBUG, " total bytes out: %10d", device_total_out);
+	logger(LOG_DEBUG, " total bytes in:  %10"PRIu64, device_total_in);
+	logger(LOG_DEBUG, " total bytes out: %10"PRIu64, device_total_out);
 }
